Move into own file, restore dependency on thumbv7em std
