+++ src/bootstrap/src/core/builder/mod.rs
@@ -873,7 +873,7 @@ impl<'a> Builder<'a> {
                 crate::ferrocene::test::GenerateTarball,
                 crate::ferrocene::test::DiffUpstream,
                 crate::ferrocene::test::certified_core_symbols::CertifiedCoreSymbols,
-                crate::ferrocene::test::FlipLink,
+                crate::ferrocene::test::flip_link::FlipLink,
                 crate::core::build_steps::toolstate::ToolStateCheck,
                 test::Tidy,
                 test::BootstrapPy,
