+++ src/bootstrap/src/ferrocene/test.rs
@@ -2,10 +2,9 @@
 // SPDX-FileCopyrightText: The Ferrocene Developers
 
 pub(crate) mod certified_core_symbols;
+pub(crate) mod flip_link;
 
-use super::tool::flip_link::PATH as FLIP_LINK_PATH;
 use crate::builder::{Builder, RunConfig, ShouldRun, Step};
-use crate::core::build_steps::compile::Std;
 use crate::core::build_steps::tool::{self, SourceType};
 use crate::core::config::TargetSelection;
 use crate::ferrocene::sign::error_when_signatures_are_ignored;
@@ -202,48 +201,6 @@ impl Step for DiffUpstream {
         builder.info("test diff-upstream");
         let mut cmd = BootstrapCommand::new("ferrocene/tools/diff-upstream/diff.py");
         cmd.arg("--hide-diff");
-        cmd.run(builder);
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub(crate) struct FlipLink {}
-
-impl Step for FlipLink {
-    type Output = ();
-    const IS_HOST: bool = true;
-
-    fn is_default_step(builder: &Builder<'_>) -> bool {
-        builder.targets.iter().any(|target| target.contains("thumbv7em"))
-    }
-
-    fn should_run(run: ShouldRun<'_>) -> ShouldRun<'_> {
-        run.path(FLIP_LINK_PATH)
-    }
-
-    fn make_run(run: RunConfig<'_>) {
-        run.builder.ensure(FlipLink {});
-    }
-
-    fn run(self, builder: &Builder<'_>) -> Self::Output {
-        let host = builder.config.host_target;
-        let compiler = builder.compiler(builder.top_stage, host);
-        builder.ensure(Std::new(compiler, host));
-
-        builder.info("Testing ferrocene/tools/flip-link");
-        let mut cmd = tool::prepare_tool_cargo(
-            builder,
-            compiler,
-            Mode::ToolTarget,
-            host,
-            Kind::Test,
-            FLIP_LINK_PATH,
-            SourceType::Submodule,
-            &[],
-        )
-        .into_cmd();
-        // Since flip-link tests the target, unsetting RUSTFLAGS prevents us from passing invalid args into the build.
-        cmd.env_remove("RUSTFLAGS");
         cmd.run(builder);
     }
 }
